Tabla terminada, con paginación
